Enable data block compression

+++ table/table_builder_computeside.cc
@@ -317,7 +317,7 @@ void TableBuilder_ComputeSide::FinishDataBlock(BlockBuilder* block, BlockHandle*
   }
   r->compressed_output.clear();
   if (r->offset < 8192){
-    printf("data content around the compression type : %p\n", block_contents.data()+block_contents.size()-4);
+    printf("data content around the compression type : %p\n", (uint64_t)block_contents.data()+block_contents.size()-4);
   }
   block->Reset_Forward();
 }
